Make MetadataMap a template (#27417)
* Make MetadataMap a template

Ultimately we'll use MetadataMap template args to specify traits for the
builtin headers they support, but initially we're just gonna ignore
them to get the syntax in place.

Necessitates moving a bunch of logic from the .cc to the .h. I tried to
do so with the minimum number of code movements... ultimately most of
this will go.

* Automated change: Fix sanity tests
